# pkg/trace/writer: use a pool for write buffers