<!--
* New contributors are highly encouraged to read our
  [CONTRIBUTING](/CONTRIBUTING.md) documentation.
* Both Contributor and Reviewer Checklists are available at https://datadoghq.dev/datadog-agent/guidelines/contributing/#pull-requests.
* The pull request:
  * Should only fix one issue or add one feature at a time.
  * Must update the test suite for the relevant functionality.
  * Should pass all status checks before being reviewed or merged.
* Commit titles should be prefixed with general area of pull request's change.
-->
### What does this PR do?
To reduce allocations, we introduce a pool to reuse serialization buffers in the trace writer.

```
name                old time/op    new time/op    delta
Serialize/large-16    1.03ms ± 5%    0.65ms ± 2%  -37.02%  (p=0.000 n=10+10)
Serialize/small-16    20.0µs ± 1%    20.0µs ± 1%     ~     (p=0.853 n=10+10)

name                old alloc/op   new alloc/op   delta
Serialize/large-16     454kB ± 2%     161kB ± 2%  -64.45%  (p=0.000 n=10+10)
Serialize/small-16    2.77kB ± 2%    1.30kB ± 5%  -53.12%  (p=0.000 n=10+10)

name                old allocs/op  new allocs/op  delta
Serialize/large-16      8.00 ± 0%      8.00 ± 0%     ~     (all equal)
Serialize/small-16      8.00 ± 0%      8.00 ± 0%     ~     (all equal)
```

### Motivation

Reduce garbage and therefore RAM usage.

### Additional Notes

The benchmark here benefits from having a large number of same-sized allocations, meaning this is probably an upper-bound of the benefit. In actuality, the benefit is likely somewhat smaller, but still worth doing.

### Possible Drawbacks / Trade-offs

<!--
* What are the possible side-effects or negative impacts of the code change?
-->

### Describe how to test/QA your changes

<!--
* Write here in detail or link to detailed instructions on how this change can
  be tested/QAd/validated, including any environment setup.
-->